+++ seqr/views/apis/locus_list_api.py
@@ -26,10 +26,11 @@ def locus_lists(request):
         Q(is_public=True) | Q(created_by=request.user)
     ).annotate(num_projects=Count('projects'))
 
-    locus_lists_json = get_json_for_locus_lists(locus_list_models, request.user, include_project_count=True)
+    locus_lists_json = get_json_for_locus_lists(locus_list_models, request.user, include_project_count=True, include_genes=True)
 
     return create_json_response({
-        'locusListsByGuid': {locus_list['locusListGuid']: locus_list for locus_list in locus_lists_json}
+        'locusListsByGuid': {locus_list['locusListGuid']: locus_list for locus_list in locus_lists_json},
+        'genesById': _get_locus_lists_genes(locus_lists_json),
     })
 
 
@@ -41,11 +42,17 @@ def locus_list_info(request, locus_list_guid):
         check_multi_project_permissions(locus_list, request.user)
 
     locus_list_json = get_json_for_locus_list(locus_list, request.user)
-    gene_ids = [item['geneId'] for item in locus_list_json['items'] if item.get('geneId')]
     return create_json_response({
         'locusListsByGuid': {locus_list_guid: locus_list_json},
-        'genesById': get_genes(gene_ids)
+        'genesById': _get_locus_lists_genes([locus_list_json]),
     })
+
+
+def _get_locus_lists_genes(locus_lists):
+    gene_ids = set()
+    for locus_list in locus_lists:
+        gene_ids.update([item['geneId'] for item in locus_list['items'] if item.get('geneId')])
+    return get_genes(gene_ids)
 
 
 @login_and_policies_required