+++ seqr/views/utils/dataset_utils.py
@@ -13,20 +13,39 @@ logger = SeqrLogger(__name__)
 SAMPLE_FIELDS_LIST = ['samples', 'samples_num_alt_1']
 
 
-def get_elasticsearch_index_samples(elasticsearch_index):
+def validate_index_metadata_and_get_elasticsearch_index_samples(elasticsearch_index, **kwargs):
     es_client = get_es_client()
 
-    index_metadata = get_index_metadata(elasticsearch_index, es_client, include_fields=True).get(elasticsearch_index)
+    all_index_metadata = get_index_metadata(elasticsearch_index, es_client, include_fields=True)
+    if elasticsearch_index in all_index_metadata:
+        index_metadata = all_index_metadata.get(elasticsearch_index)
+        validate_index_metadata(index_metadata, elasticsearch_index, **kwargs)
+        sample_field = _get_samples_field(index_metadata)
+        sample_type = index_metadata['sampleType']
+    else:
+        # Aliases return the mapping for all indices in the alias
+        metadatas = list(all_index_metadata.values())
+        sample_field = _get_samples_field(metadatas[0])
+        sample_type = metadatas[0]['sampleType']
+        for metadata in metadatas[1:]:
+            validate_index_metadata(metadata, elasticsearch_index, **kwargs)
+            if sample_field != _get_samples_field(metadata):
+                raise ValueError('Found mismatched sample fields for indices in alias')
+            if sample_type != metadata['sampleType']:
+                raise ValueError('Found mismatched sample types for indices in alias')
 
-    sample_field = next((field for field in SAMPLE_FIELDS_LIST if field in index_metadata['fields'].keys()))
     s = elasticsearch_dsl.Search(using=es_client, index=elasticsearch_index)
     s = s.params(size=0)
     s.aggs.bucket('sample_ids', elasticsearch_dsl.A('terms', field=sample_field, size=10000))
     response = s.execute()
-    return [agg['key'] for agg in response.aggregations.sample_ids.buckets], index_metadata
+    return [agg['key'] for agg in response.aggregations.sample_ids.buckets], sample_type
 
 
-def validate_index_metadata(index_metadata, project, elasticsearch_index, genome_version=None,
+def _get_samples_field(index_metadata):
+    return next((field for field in SAMPLE_FIELDS_LIST if field in index_metadata['fields'].keys()))
+
+
+def validate_index_metadata(index_metadata, elasticsearch_index, project=None, genome_version=None,
                             dataset_type=Sample.DATASET_TYPE_VARIANT_CALLS):
     metadata_fields = ['genomeVersion', 'sampleType', 'sourceFilePath']
     if any(field not in (index_metadata or {}) for field in metadata_fields):