+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # _seqr_ Changes
 
 ## dev
+* Support loading/ searching data using ES aliases
 * Added changelog